:hammer: Restructure architecture file hierarchy
