+++ src/architecture/definitions/ModuleOpening.h
