+++ src/architecture/entities/Channel.h
