+++ src/architecture/entities/Channel.hh
