+++ src/architecture/entities/Edge.h
