+++ src/architecture/entities/Edge.hh
