+++ src/architecture/entities/Membrane.h
