+++ src/architecture/entities/Membrane.hh
