+++ src/architecture/entities/Node.h
