+++ src/architecture/entities/Tank.h
