+++ src/baseSimulator.h
@@ -39,17 +39,19 @@
 #include "hybridDynamics/Scheme.h"
 #include "hybridDynamics/Naive.h"
 
-#include "architecture/Channel.h"
-#include "architecture/ChannelPosition.h"
-#include "architecture/Edge.h"
-#include "architecture/FlowRatePump.h"
-#include "architecture/Membrane.h"
-#include "architecture/Module.h"
-#include "architecture/ModuleOpening.h"
+#include "architecture/definitions/ChannelPosition.h"
+#include "architecture/definitions/ModuleOpening.h"
+
+#include "architecture/entities/Channel.h"
+#include "architecture/entities/Edge.h"
+#include "architecture/entities/FlowRatePump.h"
+#include "architecture/entities/Membrane.h"
+#include "architecture/entities/Module.h"
+#include "architecture/entities/Node.h"
+#include "architecture/entities/PressurePump.h"
+#include "architecture/entities/Tank.h"
+
 #include "architecture/Network.h"
-#include "architecture/Node.h"
-#include "architecture/PressurePump.h"
-#include "architecture/Tank.h"
 
 #include "olbProcessors/navierStokesAdvectionDiffusionCouplingPostProcessor2D.h"
 #include "olbProcessors/saturatedFluxPostProcessor2D.h"
